CSV: integrated corrections from Mr. Hudelson's sheet and checked some entries against textbook.
Calculated diff between this file and Mr. Hudelson's spreadsheet (which includes up to Lesson 17 vocabulary).
When there was a difference, I checked the textbook, corrected as needed, and then noted this in the `Note/Warning` column (hence the seeming random smattering of those notes).
Also added some words from Mr. Hudelson's sheet.
Moved `(f.)` to be with the same Arabic word / column as in the textbook.