CSV-vocab: add Anki_ID column for unique ID
This is in the form of Lesson#-v-# where v is for 'Vocabulary' anc the number is (for now), the order the word appears in the chapter.
But it only matters that this ID is unqiue and doesn't change, so if words ar emissing, or added, just make up something that doesn't contradict sense too much
e.g. 1-v-3.1 if this new word would appear between the third and fourth.
This field is just getting hashed by GenAnki so it doesn't even need to make sense, just figured I would try to make it make some sense.